backend/requirements.txt
Flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
openai==1.14.3
numpy==1.26.4
pandas==2.2.1
Pillow==10.2.0
deep-translator==1.11.4
gTTS==2.5.1
SpeechRecognition==3.10.1
pydub==0.25.1
requests==2.31.0
gunicorn==21.2.0
